remove some vs warning

+++ src/lib_ccx/ccx_decoders_isdb.h
@@ -5,7 +5,7 @@
 #include "ccx_decoders_structs.h"
 
 int isdb_set_global_time(struct lib_cc_decode *dec_ctx, uint64_t timestamp);
-int isdbsub_decode(struct lib_cc_decode *dec_ctx, const uint8_t *buf, int buf_size, struct cc_subtitle *sub);
+int isdbsub_decode(struct lib_cc_decode *dec_ctx, const uint8_t *buf, size_t buf_size, struct cc_subtitle *sub);
 void delete_isdb_caption(void *isdb_ctx);
 void *init_isdb_caption(void);
